Template/CSS
Created a template to quickly use functions in PHP for each page rather
than copy/pasting a lot of code in a lot of places.

Also, started adding comments to my code in order to provide other
people the information about what a specific section may do.

Added more to the CSS page, including Grid for easier management of
sections of the page.