Layout/Backend changes
Major changes to the code from a security standpoint.

Added more pages for an admin to add or edit a guest.

Added javascript feature to search for guest in database and either auto
pull the page up if they are the only name like the one searched, or
display the results of the search if more than one.